Update neat.js
The code has been modified to enable the use of bookmarklets through this extension (use was impossible due to the widespread use of function "chrome.tabs.update" for all protocols, even when the javascript protocol is specified in the saved bookmark). The patch code (strings 1077-1091) is not beautiful, but workable. Tested in google chrome and Yandex browser.